Move theming into default theme
- remove theming code into default theme interface and implementation
- internalize theming access
