+++ styling/src/jsMain/kotlin/dev.fritz2.styling/theme/theme.kt
@@ -222,6 +222,8 @@ interface Theme {
     val toast: ToastStyles
 
     val appFrame: AppFrameStyles
+
+    val dataTableStyles: DataTableStyles
 }
 
 /**
